Stage-0/requirements.txt
Flask==3.1.2
flask-cors==6.0.1
requests==2.32.5
python-dotenv==1.1.1